requirements.txt
pypdf>=4.2.0
Pillow>=10.0.0
reportlab>=4.0.0
openai>=1.0.0
